feat: 소식 에디터 관심사 분리 (#15)
