+++ src/components/ui/Wrapper.tsx
@@ -1,0 +1,5 @@
+import { ReactNode } from 'react';
+
+export const Wrapper = (props: { children: ReactNode; className: string }) => (
+  <div {...props}>{props.children}</div>
+);
